Added Double LinkedList
Used SingleLinkedList as a basis for implementing Double LinkedList as most (if not all) of the behaviours are the same.

Briefly started work on Implementing Graphs but spent a lot of time deciding whether or not to use C#'s dictionary for the Neighbors in GraphNode. For the time being (until I implement my own version), I will default to the built-in dictionary to store the connected nodes.

I've noticed a lot of shared behaviour (obviously) between the strictly Ordered and the Unordered SingleLinkedList. My next step is going to be to refactor this two separate objects into one class.