Stack and Nodes
Rearranged Nodes to be a separate directory, as they won't just be utilized by Lists.

Implemented and Tested Stack. There's not really much to say about the Stack, other than it was way easier than LinkedList. If I find a meaningful way to extend the Stack (or discover missing functionality), I'll return to it in the future.